+++ generative_ai/text_generation/chat_simple_example.py
